cnss2: Add support to warm reboot wlan
When unload the WLAN driver on MSM, the WiFi chip is
disabled by pulling down the WLAN EN pin. Additionally,
when the WLAN driver is reloaded, the WLAN EN pin is pulled
up, which resets the state of the WiFi chip and related
registers, initiating the normal firmware download and
initialization process.

However, on QCLinux, there is no msm_pcie_pm_control function,
so when the WLAN EN pin is pulled down, the PCIe RC is not
notified. PCIe then assumes an unexpected power loss of the device,
leading to PCIe retrain failure and inability to re-establish
the PCIe connection. The phenomenon is that the WLAN driver
cannot be reloaded successfully after being uninstalled.

In this modification, the WLAN EN pin is no longer pulled
down when unload the WLAN driver. Instead, the WiFi chip
is reset to its initial state by resetting the registers,
allowing the WLAN driver to be reloaded repeatedly on QCLinux.

Change-Id: I92877d428baaa06a3cb5117bfed830c2e6a52667
CRs-Fixed: 3945920
(cherry picked from commit 94230c938bfd9eabc13859252b24317f7ac8ba69)