cnss2: Quit from d0 wow suspend with qclinux kernel
With qclinux kernel, entering sleep mode while SAP is in a
connected state will cause the PCIe stop link operation to
timeout and the suspend will also got failed. At this time,
disconnect ref-device from SAP and try to sleep again, the
system will crash due to a timeout when accessing the PCIe
PCI_CAPABILITY_LIST register.

The reason is that to achieve D0 WoW, it is necessary to
keep the PCI device active during suspend and notify the
corresponding PCIe RC not to suspend. However, in the
current QCLinux kernel, PCIe driver not provided such
API(API in msm kernel: msm_pcie_pm_control) notify the
corresponding PCIe RC not to suspend. This leads to the PCIe
RC go to suspending, it will stop the PCI link and do host
deinit while its subordinate devices are still enabled, this
causing PCI link anomalies on the device side.

This modification exits the suspend process in the D0 WoW
scenario, preventing the RC from entering suspend and thus
avoiding PCI link anomalies.

Change-Id: I6f808d6326f1e89ad7b01f74b30fef1f734be097
CRs-Fixed: 4001768